refactor: move formatters to their own file
