+++ src/utils/index.ts
@@ -1,6 +1,5 @@
 export * from "./defaultHightlightProps";
 export * from "./exampleCode";
-export * from "./formatter";
 export * from "./parser";
 export * from "./parseObj";
 export * from "./transformers";
